Create .gitkeep
This application was paired with a non-sql data "a .txt file" that contained the names of products that were bought durring a undisclosed period of time. we can determine how many times a certain product was bought by how many times its name appears in the file. This will help a store determine what needs to be restocked and what could be considered their most popular item.

the application allows you to search for a word "capitalization sensitive" and output its frequency in the file. You are also able to simply dump the unique entries within the file and freqency of that word. You can print the unique entries of the file with a histogram being outputted as well. you can exit the program and there is a hidden 5th option to back the data as a map or a dictionary

The main logic of the switch case statement was an area that I felt particularly proud of. It is short, concise, readable and scalable. If you want to add more options to it. All you have to do is add another funtion to the list of options.

some ways that I could better my code is by encapsulating certain data and values in my inventory object. I noticed that many of my value should be considered a value that belongs to inventory.However, I decided that for prototyping purposes. Its fine where it is. Leave a note and change it later if you have to.

There was nothing in particular about this code that I thought was too challenging. I am familiar with the terrain because there was a free Harvard course that I took that covered the topic of C. So C++ wasn`t too hard to take after that.

Every skill that I developed from this course can be used for other courses. In my time as a developer, I have found languages have more in common than they do different. Any algorithms or concepts that I use for C++ can be used for Python with a few modifications to make it Python complient.

to make this program maintainable, readable, and adaptable. I tried to keep related attributes and operations withing the same object. for instance, my inventory object had to deal with my database. So, anything having to deal with database operations, I left in my inventory object. anything that had to simply deal with the main body of the program without getting into the details of the logic. I left in the main body. This allows for an easy coupling or decoupling of logic from my application and allows for easy debugging.